More Fixes

+++ dependencies/all/include/vorbis/vorbisfile.h
@@ -58,7 +58,7 @@ static int _ov_header_fseek_wrap(FILE *f,ogg_int64_t off,int whence){
 #elif defined (_WIN32)
   return _fseeki64(f,off,whence);
 #else
-  return fseek(f,off,whence);
+  return fseek(f,(long)off,whence);
 #endif
 }
